backend/requirements.txt
# Core dependencies
langchain-openai==0.3.23
langchain>=0.1.0
langgraph>=0.5.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database
asyncpg>=0.27.0
psycopg2-binary>=2.9.10

# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0

# Jupyter Lab
jupyterlab>=4.0.0
ipykernel>=6.0.0
notebook>=7.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.0.0  # Linting
mypy>=1.0.0   # Type checking 